pyproject.toml
[project]
name = "skills-matrix"
version = "0.3.0"
description = "Matrice de competences Analytics"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openpyxl>=3.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "pre-commit>=4.5.0",
]

[project.scripts]
skills-matrix = "skills_matrix.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skills_matrix"]
